# pkg/database/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.18.0

package sqlc

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Acl struct {
	PType string
	V0    string
	V1    string
	V2    string
	V3    string
	V4    string
	V5    string
}

type Role struct {
	ID        int32
	Name      string
	CreatedAt pgtype.Timestamp
	UpdatedAt pgtype.Timestamp
}

type User struct {
	ID        int32
	Name      string
	Email     string
	Password  string
	RoleID    int32
	CreatedAt pgtype.Timestamp
	UpdatedAt pgtype.Timestamp
}